Add logit integration
This adds a daemonset to the cluster.

A daemonset is a pod that runs on every node

The daemonset which we have added is filebeat. It is a single container
which forwards all docker container logs (including its own) to Logit

This changes the bootstrapping step in that it requires:
- LOGIT_API_KEY
- LOGIT_ELASTICSEARCH_HOST

These logs are shipped over TLS

@tlwr @henrytk @samcrang