Example of how to add a readiness probe
This probe gets run continually. If it exits zero the pod is marked as
ready, if it exits non-zero the pod is not marked as ready. In theory,
you could upload pipelines here from a ConfigMap, but it would re-apply
them continually throughout the life of the pod.

We change the image to henryknott/concourse because the upstream image
doesn't have fly installed.