Add basic pipeline updating mechanism
We want to store pipeline configurations in a Kubernetes-native way.

We didn't implement the pushing of the pipeline in a readiness probe because
any failure would mark the whole concourse-web pod as not ready.

We didn't use a job for this because jobs cannot be updated or re-run once
configured. They are also imperative, whereas we would prefer a declarative
way of expressing the pipeline updating process.

Therefore, we chose to add a separate container which periodically
pushes the pipelines. Running `kubectl apply` after a change to the
ConfigMap will actually cause the pipeline configuration file to be
updated on the container. This means we don't have to roll the pods
after reconfiguring the pipelines.